Wgpu 24 (#112)
~~I'm not sure how to handle `diagnostic_filter` or
`diagnostic_filter_leaf` currently. (see
https://github.com/gfx-rs/wgpu/blob/trunk/CHANGELOG.md#the-diagnostic-directive-is-now-supported-in-wgsl
for a description of what it is).~~

Diagnostic filters are just unsupported for now, didn't have time to
figure out how to get them actually working. There's a test for them but
it's marked as `#[should_panic]` due to the lack of support (also the
current expected output for it is wrong as it isn't mangled, but I need
them working in the first place to get the proper mangling so w/e).

---------

Co-authored-by: JMS55 <47158642+JMS55@users.noreply.github.com>